+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #365 "BQCoreModuleExtender.setPropertyProvider(..)": deferred property creation
 * #366 Upgrade Jackson to 2.19.4
 * #367 Upgrade JUnit 5 to 5.14.1
+* #368 Injector.isSingleton(..)
 
 ## 4.0-M1